[pytest]
1.Move the functions such as 'is_container_running(...)' to the
file /tests/common/helpers/dut_utils.py.

2.Fix the logic in the function "check_alerting_message(...)".
